extract shared code between weex/web transition modules
